refactor: move useq_widgets to new top level module
