+++ src/pymmcore_widgets/config_presets/_pixel_configuration_widget.py
@@ -32,8 +32,8 @@ from pymmcore_widgets.device_properties._device_property_table import (
 )
 from pymmcore_widgets.device_properties._device_type_filter import DeviceTypeFilters
 from pymmcore_widgets.device_properties._property_widget import PropertyWidget
-from pymmcore_widgets.useq_widgets import DataTable, DataTableWidget
-from pymmcore_widgets.useq_widgets._column_info import FloatColumn, TextColumn
+from useq_widgets import DataTable, DataTableWidget
+from useq_widgets._column_info import FloatColumn, TextColumn
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
