+++ src/pymmcore_widgets/mda/_core_z.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, Literal
 from fonticon_mdi6 import MDI6
 from pymmcore_plus import CMMCorePlus
 
-from pymmcore_widgets.useq_widgets._z import ROW_TOP_BOTTOM, Mode, ZPlanWidget
+from useq_widgets._z import ROW_TOP_BOTTOM, Mode, ZPlanWidget
 
 from ._xy_bounds import MarkVisit
