+++ src/useq_widgets/__init__.py
