+++ src/useq_widgets/_channels.py
