+++ src/useq_widgets/_checkable_tabwidget_widget.py
