+++ src/useq_widgets/_column_info.py
