+++ src/useq_widgets/_data_table.py
