+++ src/useq_widgets/_mda_sequence.py
@@ -22,12 +22,12 @@ from qtpy.QtWidgets import (
 from superqt.utils import signals_blocked
 
 import pymmcore_widgets
-from pymmcore_widgets.useq_widgets._channels import ChannelTable
-from pymmcore_widgets.useq_widgets._checkable_tabwidget_widget import CheckableTabWidget
-from pymmcore_widgets.useq_widgets._grid import GridPlanWidget
-from pymmcore_widgets.useq_widgets._positions import AF_DEFAULT_TOOLTIP, PositionTable
-from pymmcore_widgets.useq_widgets._time import TimePlanWidget
-from pymmcore_widgets.useq_widgets._z import Mode, ZPlanWidget
+from useq_widgets._channels import ChannelTable
+from useq_widgets._checkable_tabwidget_widget import CheckableTabWidget
+from useq_widgets._grid import GridPlanWidget
+from useq_widgets._positions import AF_DEFAULT_TOOLTIP, PositionTable
+from useq_widgets._time import TimePlanWidget
+from useq_widgets._z import Mode, ZPlanWidget
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
