+++ src/useq_widgets/_time.py
